requirements.txt
# UPI Analytics Platform — Python Dependencies
pandas>=2.0.0
numpy>=1.24.0
duckdb>=0.9.0
plotly>=5.18.0
streamlit>=1.28.0
loguru>=0.7.0
pyyaml>=6.0
pyarrow>=14.0.0
prophet>=1.1.5
statsmodels>=0.14.0
scikit-learn>=1.3.0
scipy>=1.11.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
gitpython>=3.1.0
openpyxl>=3.1.0
pytest>=7.4.0
httpx>=0.25.0
ruff>=0.1.0
